assert_is_data.frame {assertive.types}	R Documentation

Is the input is a data.frame?

Description

Is the input is a data.frame?

Usage

assert_is_data.frame(x, severity = getOption("assertive.severity", "stop"))

is_data.frame(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_data.frame wraps is.data.frame, providing more information on failure. assert_is_data.frame returns nothing but throws an error if is_data.frame returns FALSE.

See Also

is.data.frame.

Examples

assert_is_data.frame(data.frame())
assert_is_data.frame(datasets::CO2)

[Package assertive.types version 0.0-3 Index]