assert_is_formula {assertive.types}	R Documentation

Is the input a formula?

Description

Checks to see if the input is a formula.

Usage

assert_is_formula(x, severity = getOption("assertive.severity", "stop"))

assert_is_one_sided_formula(x, severity = getOption("assertive.severity",
  "stop"))

assert_is_two_sided_formula(x, severity = getOption("assertive.severity",
  "stop"))

is_formula(x, .xname = get_name_in_parent(x))

is_one_sided_formula(x, .xname = get_name_in_parent(x))

is_two_sided_formula(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

The is_* functions return TRUE when the input is a formula. The assert_* functions return nothing but throw an error if the corresponding is_* function returns FALSE.

See Also

is_environment and is_language

Examples

is_one_sided_formula(~ x)
is_two_sided_formula(y ~ x)

[Package assertive.types version 0.0-3 Index]