assert_is_function {assertive.types}	R Documentation

Is the input a function?

Description

Checks to see if the input is a function.

Usage

assert_is_function(x, severity = getOption("assertive.severity", "stop"))

assert_is_primitive(x, severity = getOption("assertive.severity", "stop"))

assert_is_stepfun(x, severity = getOption("assertive.severity", "stop"))

is_function(x, .xname = get_name_in_parent(x))

is_primitive(x, .xname = get_name_in_parent(x))

is_stepfun(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_function, is_primitive and is_stepfun wrap is.function, is.primitive and is.stepfun repsectively, providing more information on failure. The assert_* functions return nothing but throw an error if the corresponding is_* function returns FALSE.

See Also

is.function.

Examples

assert_is_function(sqrt)
assert_is_function(function(){})

[Package assertive.types version 0.0-3 Index]