assert_is_inherited_from {assertive.types}	R Documentation

Does the object inherit from some class?

Description

Checks to see if an object is inherited from any of the specified classes.

Usage

assert_is_inherited_from(x, classes,
  severity = getOption("assertive.severity", "stop"))

is_inherited_from(x, classes, .xname = get_name_in_parent(x))

Arguments

x
Any R variable.

classes
A character vector of classes.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

TRUE if x inherits from at least one of the classes, as determined by inherits.

See Also

inherits, is, is2

Examples

x <- structure(1:5, class = c("foo", "bar"))
assert_is_inherited_from(x, c("foo", "baz"))
assertive.base::dont_stop(assert_is_inherited_from(x, c("Foo", "baz")))

[Package assertive.types version 0.0-3 Index]