assert_is_an_integer {assertive.types}	R Documentation

Is the input an integer?

Description

Checks to see if the input is an integer.

Usage

assert_is_an_integer(x, severity = getOption("assertive.severity", "stop"))

assert_is_integer(x, severity = getOption("assertive.severity", "stop"))

is_an_integer(x, .xname = get_name_in_parent(x))

is_integer(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_integer wraps is.integer, providing more information on failure. is_an_integer returns TRUE if the input is an integer and scalar. The assert_* functions return nothing but throw an error if the corresponding is_* function returns FALSE.

See Also

is.integer and is_scalar.

Examples

assert_is_integer(1:10)
assert_is_an_integer(99L)
#These examples should fail.
assertive.base::dont_stop(assert_is_integer(c(1, 2, 3)))
assertive.base::dont_stop(assert_is_an_integer(1:10))
assertive.base::dont_stop(assert_is_an_integer(integer()))

[Package assertive.types version 0.0-3 Index]