assert_is_leaf {assertive.types}	R Documentation

Is the input a (dendrogram) leaf?

Description

Checks to see if the input is a (dendrogram) leaf.

Usage

assert_is_leaf(x, severity = getOption("assertive.severity", "stop"))

is_leaf(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_leaf reimplements is.leaf, providing more information on failure.

See Also

dendrogram.

[Package assertive.types version 0.0-3 Index]