assert_is_a_bool {assertive.types}	R Documentation

Is the input logical?

Description

Checks to see if the input is logical.

Usage

assert_is_a_bool(x, severity = getOption("assertive.severity", "stop"))

assert_is_logical(x, severity = getOption("assertive.severity", "stop"))

is_a_bool(x, .xname = get_name_in_parent(x))

is_logical(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_logical wraps is.logical, providing more information on failure. is_a_bool returns TRUE if the input is logical and scalar. The assert_* functions return nothing but throw an error if the corresponding is_* function returns FALSE.

See Also

is.logical and is_scalar.

Examples

assert_is_logical(runif(10) > 0.5)
assert_is_a_bool(TRUE)
assert_is_a_bool(NA)
#These examples should fail.
assertive.base::dont_stop(assert_is_logical(1))
assertive.base::dont_stop(assert_is_a_bool(c(TRUE, FALSE)))
assertive.base::dont_stop(assert_is_a_bool(logical()))

[Package assertive.types version 0.0-3 Index]