assert_is_qr {assertive.types}	R Documentation

Is the input a QR decomposition of a matrix?

Description

Checks to see if the input is a QR decomposition of a matrix.

Usage

assert_is_qr(x, severity = getOption("assertive.severity", "stop"))

is_qr(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_qr wraps is.qr, providing more information on failure. assert_is_qr returns nothing but throws an error if is_qr returns FALSE.

See Also

is.qr.

Examples

assert_is_qr(qr(matrix(rnorm(25), nrow = 5)))

[Package assertive.types version 0.0-3 Index]